requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-core==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.8
pgvector==0.2.5
supabase==2.4.3
postgrest==2.27.0
psycopg[binary]==3.1.19
pydantic==2.7.3
pydantic-settings==2.4.0
httpx==0.27.0
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
openai==1.37.1
pdfminer.six==20231228
python-docx==0.8.11
